==> notebooks/step0_dataset_exploration.ipynb <==
import pandas as pd

df = pd.read_csv("strix.csv")

pd.set_option('display.max_colwidth', None)  # Show all columns
len(df)

# Stoic ideology filter

# Keywords related to Stoicism
stoic_keywords = ["stoicism", "stoic"]

# Filter rows where any of the keywords appear in 'category', 'question', or 'answer'
stoic_df = df[df.apply(lambda row: any(keyword in str(row).lower() for keyword in stoic_keywords), axis=1)]
len(stoic_df)

category_counts = stoic_df.groupby('category').size()
# Sort by count in descending order
top_10_categories = category_counts.sort_values(ascending=False).head(15)
top_10_categories

stoic_categories = ["stoicism","epictetus","seneca","zeno","justus-lipsius","marcus-aurelius"]
stoic_df_authors = df[df["category"].isin(stoic_categories)]
len(stoic_df_authors)

stoic_union_df = pd.concat([stoic_df, stoic_df_authors]).drop_duplicates()
stoic_union_df["ideology"] = "stoicism"
len(stoic_union_df)

stoic_union_df.head(3)

# Zen Budhism

# Keywords related to Stoicism
zen_keywords = ["zen", "zen-buddhism","zen buddhism"]

# Filter rows where any of the keywords appear as complete words in 'category', 'question', or 'answer'
zen_df = df[df.apply(lambda row: any(
    pd.Series(row).str.contains(r'\b' + keyword + r'\b', case=False, regex=True).any() for keyword in zen_keywords
), axis=1)]

# Number of matching rows
print(f'Number of matching rows: {len(zen_df)}')

category_counts = zen_df.groupby('category').size()

# Sort by count in descending order
top_10_categories = category_counts.sort_values(ascending=False).head(10)
top_10_categories

zen_categories = ["japanese-zen","kyoto-school"]
zen_authors = df[df["category"].isin(zen_categories)]
len(zen_authors)

zen_union_df = pd.concat([zen_df, zen_authors]).drop_duplicates()
zen_union_df["ideology"] = "zen-buddhism"
len(zen_union_df)

zen_union_df.head(3)

# Union stoicism and zen-buddhism QnA

stoic_zen_union_df = pd.concat([stoic_union_df, zen_union_df]).drop_duplicates()
len(stoic_zen_union_df)

stoic_zen_union_df.head(3)

stoic_zen_union_df.to_csv("stoic_zen_document.csv",index = False)
